improve: docs & algos
