+++ mkdocs.yml
@@ -133,8 +133,8 @@ plugins:
             show_root_heading: false
             show_root_full_path: true
             show_signature_annotations: true
-            show_symbol_type_heading: true
-#            show_symbol_type_toc: true
             signature_crossrefs: true
             show_root_members_full_path: true
-            show_root_toc_entry: false
+            show_root_toc_entry: true
+            show_symbol_type_heading: true
+            show_symbol_type_toc: true
