+++ pyproject.toml
@@ -59,12 +59,12 @@ pytest = "^7.1.3"
 pytest-cov = "^4.0.0"
 
 [tool.poetry.group.docs.dependencies]
-mkdocs = "^1.4.0"
-mkdocs-material = "^8.5.3"
-mkdocs-material-extensions = "^1.0.3"
-mkdocs-bibtex = "^2.8.10"
-mkdocstrings-python = "^1.10.3"
-mkdocs-autorefs = "^1.0.1"
+mkdocs = ">=1.5.0"
+mkdocs-material = ">=9.5.0"
+mkdocs-material-extensions = ">=1.0.3"
+mkdocs-bibtex = ">=2.8.10"
+mkdocstrings-python = ">=1.10.3"
+mkdocs-autorefs = ">=1.0.1"
 
 
 [[tool.poetry.source]]
