+++ toyml/clustering/__init__.py
@@ -1,5 +1,11 @@
 from .agnes import Agnes
+from .bisect_kmeans import BisectingKmeans
 from .dbscan import DbScan
-from .kmeans import BisectingKmeans, Kmeans, KmeansPlus
+from .kmeans import Kmeans
 
-__all__ = ["Kmeans", "BisectingKmeans", "KmeansPlus", "DbScan", "Agnes"]
+__all__ = [
+    "Kmeans",
+    "BisectingKmeans",
+    "DbScan",
+    "Agnes",
+]
