--- toyml/clustering/kmeans/__init__.py
@@ -1,5 +1,0 @@
-from .bisecting import BisectingKmeans
-from .plus import KmeansPlus
-from .simple import Kmeans
-
-__all__ = ["Kmeans", "BisectingKmeans", "KmeansPlus"]
